// Haly.WebApp/ClientApp/generated/haly/models/UpdatePlaylistDetailsRequestBody.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Haly API
 * An ASP.NET Core Web API that adds quality of life improvements to Spotify and helps with music exploration
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdatePlaylistDetailsRequestBody
 */
export interface UpdatePlaylistDetailsRequestBody {
    /**
     * 
     * @type {string}
     * @memberof UpdatePlaylistDetailsRequestBody
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof UpdatePlaylistDetailsRequestBody
     */
    description: string;
}

/**
 * Check if a given object implements the UpdatePlaylistDetailsRequestBody interface.
 */
export function instanceOfUpdatePlaylistDetailsRequestBody(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "description" in value;

    return isInstance;
}

export function UpdatePlaylistDetailsRequestBodyFromJSON(json: any): UpdatePlaylistDetailsRequestBody {
    return UpdatePlaylistDetailsRequestBodyFromJSONTyped(json, false);
}

export function UpdatePlaylistDetailsRequestBodyFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdatePlaylistDetailsRequestBody {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'description': json['description'],
    };
}

export function UpdatePlaylistDetailsRequestBodyToJSON(value?: UpdatePlaylistDetailsRequestBody | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'description': value.description,
    };
}